move tests into a separate file
