+++ apollo-router/src/axum_factory/mod.rs
@@ -3,6 +3,8 @@ mod axum_http_server_factory;
 mod handlers;
 mod listeners;
 mod utils;
+#[cfg(test)]
+mod tests;
 
 pub(crate) use axum_http_server_factory::make_axum_router;
 pub(crate) use axum_http_server_factory::AxumHttpServerFactory;
